+++ src/shared/ui/SelectGroup.tsx
@@ -1,0 +1,3 @@
+import * as SelectPrimitive from '@radix-ui/react-select';
+
+export const SelectGroup = SelectPrimitive.Group;
