+++ src/shared/ui/SelectValue.tsx
@@ -1,0 +1,3 @@
+import * as SelectPrimitive from '@radix-ui/react-select';
+
+export const SelectValue = SelectPrimitive.Value;
